libc: replaced defines with enums/constants.
Some are still required, because "const" in C does not
mean the same as in C++. In C, it's a readonly variable
and thus can't be used as a constant expression (e.g.
to specify an static array dimension).

+++ source/include/vterm/vtctrl.h
@@ -120,7 +120,7 @@ struct sVTerm {
 	/* the pid of the shell for ctrl+c notifications */
 	pid_t shellPid;
 	/* the escape-state */
-	int escapePos;
+	size_t escapePos;
 	char escapeBuf[MAX_ESCC_LENGTH];
 	/* readline-buffer */
 	size_t rlStartCol;
